requirements.txt
streamlit
google-genai
python-dotenv
speechrecognition
streamlit-mic-recorder
pymongo